+++ kinto_http/exceptions.py
@@ -32,3 +32,10 @@ class BackoffException(KintoException):
     def __init__(self, message, backoff, exception=None):
         self.backoff = backoff
         super().__init__(message, exception)
+
+
+class KintoBatchException(KintoException):
+    def __init__(self, exceptions, results):
+        self.message = "\n".join([str(e) for e in exceptions])
+        self.exceptions = exceptions
+        self.results = results